+++ src/days/day15.rs
@@ -87,10 +87,10 @@ fn part1(mut warehouse: Grid<u8>, movements: &[Direction]) -> String {
         }
     }
 
-    let total_gps_coordinates: usize = warehouse
+    let total_gps_coordinates: isize = warehouse
         .iter_with_coords()
         .filter(|(_, ch)| **ch == b'O')
-        .map(|(coord, _)| coord.r as usize * 100 + coord.c as usize)
+        .map(|(coord, _)| coord.r * 100 + coord.c)
         .sum();
     format!("{total_gps_coordinates}")
 }
@@ -192,10 +192,10 @@ fn part2(mut warehouse: Grid<u8>, movements: &[Direction]) -> String {
         }
     }
 
-    let total_gps_coordinates: usize = warehouse
+    let total_gps_coordinates: isize = warehouse
         .iter_with_coords()
         .filter(|(_, ch)| **ch == b'[')
-        .map(|(coord, _)| coord.r as usize * 100 + coord.c as usize)
+        .map(|(coord, _)| coord.r * 100 + coord.c)
         .sum();
     format!("{total_gps_coordinates}")
 }
